chore: rename other changed files to match refactor
